# a7/a7_openapi.yaml
openapi: "3.0.0"
info:
  version: "1.0"
  title: "LBAW Online Auctions Web API"
  description: "Web Resources Specification (A7) for Online Auctions"
servers:
  - url: http://lbaw.fe.up.pt
    description: "production server"
...
paths:
...
  /api/item:
    get:
      operationId: R202
      summary: "R202: Search Works API"
      description: "Searches for items and returns the results as JSON. Access: PUB."

      tags:
        - "M02: Items"

      parameters:
        - in: query
          name: query
          description: String to use for full-text search
          schema:
            type: string
          required: false
        - in: query
          name: name
          description: Name of the item
          schema:
            type: string
          required: false
        - in: query
          name: description
          description: Description of the auction
          schema:
            type: string
          required: false

      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    description:
                      type: string
                example:
                  - id: 1
                    name: Rihanna Vinyl
                    description: Good pop music album.
                  - id: 2
                    name: Mr. Bean Movie DVD
                    description: Very funny movie.
...
  /auction/{id}:
    get:
      operationId: R401
      summary: "R401: View Auction Page"
      description: "Show the individual auction page. Access: PUB"
      tags:
        - "M04: Auctions"

      parameters:
        - in: path
          name: id_auction
          schema:
            type: integer
          required: true

      responses:
        "200":
          description: "Ok. Show Auction Page UI"
...